pyproject.toml
[tool.poetry]
name = "FlaskAPIdemo"
version = "0.1.0"
description = ""
authors = ["manojbaishya <manojbaishya1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
Jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^5.2"
flake8 = "^3.9.1"
flake8-flask = "^0.9.3"
pylint-flask-sqlalchemy = "^0.2.0"
ipython = "^7.22.0"
Faker = "^8.1.1"
pyproject-flake8 = "^0.0.1-alpha.2"
tomlkit = "^0.7.0"

[tool.pyright]
exclude = [".venv"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
